Refactor TTS functionality: Moved BrowserTTS component and updated imports
